Moved Upload extensions into a separate Swift file.
